+++ cvat/apps/engine/media_io/frame_provider.py
