+++ cvat/apps/quality_control/quality_reports.py
@@ -40,7 +40,7 @@ from cvat.apps.dataset_manager.bindings import (
 from cvat.apps.dataset_manager.formats.registry import dm_env
 from cvat.apps.dataset_manager.task import JobAnnotation
 from cvat.apps.engine.filters import JsonLogicFilter
-from cvat.apps.engine.frame_provider import TaskFrameProvider
+from cvat.apps.engine.media_io.frame_provider import TaskFrameProvider
 from cvat.apps.engine.model_utils import bulk_create
 from cvat.apps.engine.models import (
     DimensionType,
